BLD: [CPP] Split out bindings from implementation
