+++ cmake/MMU_ReleaseFlags.cmake
@@ -1,0 +1,19 @@
+INCLUDE(CheckCXXCompilerFlag)
+SET(MMU_RELEASE_FLAGS "")
+CHECK_CXX_COMPILER_FLAG("-march=native" COMPILER_SUPPORTS_MARCH_NATIVE)
+IF(COMPILER_SUPPORTS_MARCH_NATIVE)
+    SET(MMU_RELEASE_FLAGS "${MMU_RELEASE_FLAGS} -march=native")
+ENDIF()
+CHECK_CXX_COMPILER_FLAG("-mtune=native" COMPILER_SUPPORTS_MTUNE_NATIVE)
+IF(COMPILER_SUPPORTS_MTUNE_NATIVE)
+    SET(MMU_RELEASE_FLAGS "${MMU_RELEASE_FLAGS} -mtune=native")
+ENDIF()
+CHECK_CXX_COMPILER_FLAG("-ftree-vectorize" COMPILER_SUPPORTS_FTREE)
+IF(COMPILER_SUPPORTS_FTREE)
+    SET(MMU_RELEASE_FLAGS "${MMU_RELEASE_FLAGS} -ftree-vectorize")
+ENDIF()
+CHECK_CXX_COMPILER_FLAG("-mavx2" COMPILER_SUPPORTS_MAVX2)
+IF(COMPILER_SUPPORTS_MAVX2)
+    SET(MMU_RELEASE_FLAGS "${MMU_RELEASE_FLAGS} -mavx2")
+ENDIF()
+
