+++ cpp/bindings/confusion_matrix.cpp
@@ -1,29 +1,7 @@
-/* metrics.hpp -- Implementation of binary classification metrics
+/* confusion_matrix_bindings.cpp -- Python bindings for confusion_matrix.hpp metrics
  * Copyright 2021 Ralph Urlus
  */
-#ifndef MMU_METRICS_CONFUSION_MATRIX_HPP_
-#define MMU_METRICS_CONFUSION_MATRIX_HPP_
-
-/* TODO *
- *
- * - Add function over runs for yhat
- * - Add function over runs for proba with single threshold
- * - Add function over runs for proba over multiple thresholds
- * - Add support for int confusion_matrices
- *
- * TODO */
-
-#include <pybind11/pybind11.h>
-#include <pybind11/numpy.h>
-#include <pybind11/stl.h>
-
-#include <cmath>
-#include <string>
-#include <limits>
-#include <cinttypes>
-#include <type_traits>
-
-#include "commons/utils.hpp"
+#include "confusion_matrix.hpp"
 
 namespace py = pybind11;
 
@@ -40,67 +18,7 @@ namespace py = pybind11;
  */
 
 namespace mmu {
-namespace details {
-
-// --- from yhat ---
-template<class T, std::enable_if_t<std::is_integral<T>::value, int> = 0>
-inline void confusion_matrix(const size_t n_obs, T* y, T* yhat, int64_t* const conf_mat) {
-    for (size_t i = 0; i < n_obs; i++) {
-        conf_mat[static_cast<bool>(*y) * 2 + static_cast<bool>(*yhat)]++; yhat++; y++;
-    }
-}
-
-template<class T, std::enable_if_t<std::is_floating_point<T>::value, int> = 1>
-inline void confusion_matrix(const size_t n_obs, T* y,  T* yhat, int64_t* const conf_mat) {
-    for (size_t i = 0; i < n_obs; i++) {
-        conf_mat[(*y > 1e-12) * 2 + (*yhat > 1e-12)]++; yhat++; y++;
-    }
-}
-
-// --- from proba and threshold ---
-template<class T, typename A, std::enable_if_t<std::is_integral<T>::value, int> = 0>
-inline void confusion_matrix(
-    const size_t n_obs, T* y, A* proba, const A threshold, int64_t* const conf_mat
-) {
-    for (size_t i = 0; i < n_obs; i++) {
-        conf_mat[static_cast<bool>(*y) * 2 + (*proba > threshold)]++; proba++; y++;
-    }
-}
-
-template<class T, typename A, std::enable_if_t<std::is_floating_point<T>::value, int> = 1>
-inline void confusion_matrix(
-    const size_t n_obs, T* y, A* proba, const A threshold, int64_t* const conf_mat
-) {
-    for (size_t i = 0; i < n_obs; i++) {
-        conf_mat[(*y > 1e-12) * 2 + (*proba > threshold)]++; proba++; y++;
-    }
-}
-
-}  // namespace details
-
-
 namespace bindings {
-
-template <typename T>
-py::array_t<int64_t> confusion_matrix(const py::array_t<T>& y, const py::array_t<T>& yhat) {
-    details::check_1d_soft(y, "y");
-    details::check_contiguous(y, "y");
-    details::check_1d_soft(yhat, "yhat");
-    details::check_contiguous(yhat, "yhat");
-    details::check_equal_length(y, yhat);
-
-    // note memory is uninitialised
-    auto conf_mat = py::array_t<int64_t>({2, 2}, {16, 8});
-    int64_t* cm_ptr = reinterpret_cast<int64_t*>(conf_mat.request().ptr);
-    // initialise the memory
-    memset(cm_ptr, 0, sizeof(int64_t) * 4);
-
-    auto y_ptr = reinterpret_cast<T*>(y.request().ptr);
-    auto yhat_ptr = reinterpret_cast<T*>(yhat.request().ptr);
-    details::confusion_matrix<T>(y.size(), y_ptr, yhat_ptr, cm_ptr);
-
-    return conf_mat;
-}
 
 void bind_confusion_matrix(py::module &m) {
     m.def(
@@ -157,29 +75,6 @@ void bind_confusion_matrix(py::module &m) {
         py::arg("y"),
         py::arg("yhat")
     );
-}
-
-template <typename T>
-py::array_t<int64_t> confusion_matrix(
-    const py::array_t<T>& y, const py::array_t<double>& proba, const double threshold
-) {
-    details::check_1d_soft(y, "y");
-    details::check_contiguous(y, "y");
-    details::check_1d_soft(proba, "proba");
-    details::check_contiguous(proba, "proba");
-    details::check_equal_length(y, proba);
-
-    // note memory is uninitialised
-    auto conf_mat = py::array_t<int64_t>({2, 2}, {16, 8});
-    int64_t* cm_ptr = reinterpret_cast<int64_t*>(conf_mat.request().ptr);
-    // initialise the memory
-    memset(cm_ptr, 0, sizeof(int64_t) * 4);
-
-    auto y_ptr = reinterpret_cast<T*>(y.request().ptr);
-    auto proba_ptr = reinterpret_cast<double*>(proba.request().ptr);
-    details::confusion_matrix<T>(y.size(), y_ptr, proba_ptr, threshold, cm_ptr);
-
-    return conf_mat;
 }
 
 void bind_confusion_matrix_proba(py::module &m) {
@@ -248,5 +143,3 @@ void bind_confusion_matrix_proba(py::module &m) {
 
 }  // namespace bindings
 }  // namespace mmu
-
-#endif  // MMU_METRICS_CONFUSION_MATRIX_HPP_
